refact: move SectionMain to pages
